requirements.txt
flask
spacy
easyocr
opencv-python
torch  
python-decouple